move chunks out of mount
