+++ zerostash-fuse/src/lib.rs
@@ -1,5 +1,6 @@
 #![deny(unused_crate_dependencies)]
 pub mod mount;
+pub mod chunks;
 
 #[cfg(test)]
 use criterion as _;
